Extract statements_explain_describe.go
